backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.1
django-admin-volt==1.0.17
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
msgpack==1.1.0
pillow==11.2.1
PyJWT==2.9.0
python-decouple==3.8
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.13.2
